Provide a way to disable logging #273
This adds a `verbose` option to LS configuration. The default value is `verbose=true` enables the messages to be logged
using `connection.console.log(message)`,  while `verbose=false` disables any messages to be logged.
This doesn't change the logging of any other types, f.i. when `connection.console.warn(message)`  is used the message will be
sent to console regardless of value of `verbose` option.

Signed-off-by: Victor Rubezhny <vrubezhny@redhat.com>